Allow for witZaleniums()'s parameter to be overwritten individually.
While keeping the other parameters on default.
There seems to be not option for named method parameters.

Before this commit a call of witZaleniums(zaleniumVideoDir: 'something')
would result in only the zaleniumVideoDir being set and all
default values being null. This is no surprise, because the whole map
is overwritten.