#4 Support automatic Maven-settings.xml, support truststore
# withMavenSettings

The function `withMavenSettings() { -> settingsXml }` provides an easy-
to-use functionality which writes a settings file and cleans it up
afterwards.

# truststore

Sharing a truststore is a fickle thing because it contains crucial
certificates. Once a build job finishes it is unnecessarily hard to
remove truststore files. For this reason `copy()` and `use()` exist
so that the files are removed on different Jenkins nodes within the
Job build workspace.