#7 Improve hub status check
The Selenium documentation tells us (if we want to check if hub
is available) that we should consult the `value.ready` object rather
than waiting for `status` to be zero.

The status indicates the plain availability of the hub while `ready`
indicates if the hub has actual capacity to start any tests, that is
if any are readily available.

Signed-off-by: Philipp Pixel <pixeloffice@philipppixel.de>